Export UMD module
The changes how the library is bundled. Instead of just running
it through babel and requiring a commonjs2 like bundler on the user
side, it will export as an UMD module instead. This means most bundlers,
as well as global definition will work out of the box.

The drawback is a slight increase in bundle size. But the net change will be
less since it no longer bundles react or react-dom (:doh:).

Fixes #5